pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-platform-ai-orchestration"
version = "1.0.0"
description = "Strategic Multi-Platform AI Orchestration Implementation with Distributed Autonomous Capabilities"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "AI Orchestration Team"}
]
keywords = ["ai", "orchestration", "distributed", "autonomous", "strategic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "typer[all]>=0.9.0",
    "openai>=1.0.0",
    "google-generativeai>=0.5.0",
    "anthropic>=0.20.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[project.urls]
Homepage = "https://github.com/poisontr33s/Multi-PlatformAIOrchestrationImplementation"
Repository = "https://github.com/poisontr33s/Multi-PlatformAIOrchestrationImplementation"
Issues = "https://github.com/poisontr33s/Multi-PlatformAIOrchestrationImplementation/issues"

[project.scripts]
ai-orchestrator = "src.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
norecursedirs = ["archive", "static", ".venv"]
pythonpath = ["src"]
